requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
openai==1.3.7
PyPDF2==3.0.1
boto3==1.34.0
requests==2.31.0
pydantic==2.5.0
alembic==1.13.0 